+++ Source/IGListSingleSectionController.h
@@ -53,6 +53,18 @@ NS_SWIFT_NAME(ListSingleSectionControllerDelegate)
 - (void)didSelectSectionController:(IGListSingleSectionController *)sectionController
                         withObject:(id)object;
 
+/**
+ Tells the delegate that the section controller was deselected.
+
+ @param sectionController The section controller that was deselected.
+ @param object The model for the given section.
+
+ @note Method is `@optional` until the 4.0.0 release where it will become required.
+ */
+@optional
+- (void)didDeselectSectionController:(IGListSingleSectionController *)sectionController
+                          withObject:(id)object;
+
 @end
 
 /**